fatJar for executing
Added the Shadow Gradle plugin to create a fatJar for distribution

Also, it turns out Alex was correct when he separated main() from the start() method. IDK why this is the case, but not having it structured this way causes compilation errors with JavaFX and it took me all morning to figure out and it was very frustrating. :(

I also fixed some reference errors in the PDFConverterTest file so that I could compile.